✨ feat: add food feats 1. update sold out in table 2. conform modal before delete 3. add MySwitch

+++ src/stores/foods.js
@@ -1,7 +1,9 @@
 import { defineStore } from 'pinia'
 import api from '@/utils/api'
 import { useShopStore } from './shop'
-import { debug } from '../utils/swal'
+import { debug, ng } from '../utils/swal'
+import { useGlobalStore } from './global'
+import { useRouter } from 'vue-router'
 const foodForEdited = {
   foodTypeId: null,
   name: '',
@@ -22,8 +24,7 @@ export const useFoodStore = defineStore('food', {
     },
     foods: [],
     foodTypes: [],
-    activeFoodTypeIndex: null,
-    activeFoodId: null,
+    activeFoodTypeId: null,
     foodForEdited: {
       foodTypeId: null,
       name: '',
@@ -34,9 +35,6 @@ export const useFoodStore = defineStore('food', {
     }
   }),
   getters: {
-    activeFood (state) {
-      return state.foods.find(e => e.id === state.activeFoodId)
-    },
     skip (state) {
       const { page, take } = state.pagination
       return take * (page - 1)
@@ -60,13 +58,18 @@ export const useFoodStore = defineStore('food', {
       }
     },
     async fetchFoods (shopId, foodTypeId) {
-      const { take, orderBy, orderType } = this.pagination
-      const query = Object.entries({ take, skip: this.skip, orderBy, orderType })
-        .map(([k, v]) => `${k}=${v}`).join('&')
-      const url = `shops/${shopId}/foodTypes/${foodTypeId}/foods?${query}`
-      const { foods, total } = await api(url)
-      this.pagination.max = Math.ceil(total / this.pagination.take)
-      this.foods = foods
+      try {
+        const { take, orderBy, orderType } = this.pagination
+        const query = Object.entries({ take, skip: this.skip, orderBy, orderType })
+          .map(([k, v]) => `${k}=${v}`).join('&')
+        const url = `shops/${shopId}/foodTypes/${foodTypeId}/foods?${query}`
+        const { foods, total } = await api(url)
+        this.activeFoodTypeId = +foodTypeId
+        this.pagination.max = Math.ceil(total / this.pagination.take)
+        this.foods = foods
+      } catch (err) {
+        ng('no fetchFoods')
+      }
     },
     async fetchFood (foodId) {
       const shopId = useShopStore().shop.id
@@ -74,31 +77,57 @@ export const useFoodStore = defineStore('food', {
       this.foodForEdited = { ...food }
     },
     async createFood (payload) {
+      const globalStore = useGlobalStore()
+      if (globalStore.isLoading) return
       const shopId = useShopStore().shop.id
       if (shopId == null) throw new Error('Create food, need shopId')
       payload = { ...payload, shopId }
       const food = await api.post(`shops/${shopId}/foods`, payload)
       return food
     },
+    async updateFood (payload) {
+      const globalStore = useGlobalStore()
+      if (globalStore.isLoading) return
+      const {
+        id: foodId,
+        name, info, price, isSoldOut, imgId
+      } = payload
+      const apiPayload = { foodId, name, info, price, isSoldOut, imgId }
+      const shopId = useShopStore().shop.id
+      if (shopId == null) throw new Error('Update food, need shopId')
+      const food = await api.put(`shops/${shopId}/foods/${foodId}`, apiPayload)
+      return food
+    },
     async delete (foodId) {
+      const globalStore = useGlobalStore()
+      if (globalStore.isLoading) return
       try {
         const shopId = useShopStore().shop.id
-        const res = api.delete(`shops/${shopId}/foods/${foodId}`)
-        this.foods = this.foods.filter(e => e.id !== foodId)
+        await api.delete(`shops/${shopId}/foods/${foodId}`)
+        await this.fetchFoods(shopId, this.activeFoodTypeId)
+        // this.foods = this.foods.filter(e => e.id !== foodId)
       } catch (err) {
-        debug(err)
+        console.log(err)
+        ng('delete error')
       }
     },
     async submitFood (type) {
       if (type === 'create') {
-        console.log('create')
-        console.log(this.foodForEdited)
-        return
+        await this.createFood(this.foodForEdited)
       }
       if (type === 'update') {
-        console.log('update')
-        console.log(this.foodForEdited)
+        await this.updateFood(this.foodForEdited)
       }
+    },
+    async updateSoldOut (foodId, value) {
+      if (foodId == null || foodId === '') throw new Error('updateSoldOut:need foodId')
+      const shopId = useShopStore().shop.id
+      if (shopId == null) throw new Error('updateSoldOut: need shopId')
+      const { food } = await api.patch(`shops/${shopId}/foods/${foodId}`, { colName: 'isSoldOut', value })
+      if (!food) throw new Error('updateSoldOut')
+
+      const target = this.foods.find(f => +f.id === +foodId)
+      target.isSoldOut = food.isSoldOut
     }
   }
 })
